use float instead of double in jura::TrackMeter

+++ Libraries/RobsJuceModules/jura_processors/analyzers/jura_TrackMeter.h
@@ -77,7 +77,7 @@ protected:
   MeteringDisplay *leftLevelMeter, *rightLevelMeter, *midLevelMeter, *sideLevelMeter, 
     *correlationMeter;
 
-  double rangeMin, rangeMax;
+  float rangeMin, rangeMax;
 
   juce_UseDebuggingNewOperator;
 };
